make_backups.sh: keep three source snapshots
This prevents breaking the chain of snapshots in the event of a reboot in the
middle of a send.  When keeping only two snapshots, the last common parent is
deleted by btrfs-snap, so that is not enough.

This is not the most elegant way of dealing with this, but it's the least
invasive.