gen_kernel: append a count to the initramfs, too
When a kernel image has a count ("-c<N>") appended to its version, its
corresponding initramfs must have one, too, in order for grub2-mkconfig to add
it to the menu entry.

I found out about this because diefledermaus was failing to boot.  Eventually I
tracked it down to a missing initramfs.