Add a script that generates a Gentoo nspawn container
This commit adds a script that generates a Gentoo container suitable for
use with systemd-nspawn.  It fetches the most recent amd64-systemd
stage3 tarball and automatically runs some commands from within the
container to prepare it; it also generates a *.nspawn unit file on the
host for mounting some of its directories read-only into the container.
The container produced in this way can be used to test ebuilds in a
stable amd64 environment.

Right now it is concretely targeted at testing asciidoc (since I am
proxy maintainer for it), but could probably be generalised with a bit
of work (e.g., by directly copying a prepared /etc/portage directory
into the container instead of hard-coding commands to be run).

In any case, this should make it easier for me to (re-)generate testing
containers on any of my Gentoo machines.