init_gentoo_container.sh: use config directories
Rework init_gentoo_container.sh to use a configuration repository with
one directory per container (which is, however, optional).  This is to
make the script more suitable for creating container images for testing
other packages.

The way this works now is as follows (with ${CONFIG_DIR} as the
container's configuration directory):

1.) The container image is downloaded.
2.) An optional systemd nswap unit is installed on the host.
3.) systemd-nspawn runs a series of commands in the container (with
  ${CONFIG_DIR} bind mounted therein as /root/container.conf/):
    3a.) Files in filesystem/ are installed into the root file system.
    3b.) The portage tree is initialised via emerge-webrsync followed by
    "emerge --sync".
    3c.) All files in scripts.preinst/ ending in .sh are executed.
    3d.) A @world update is performed.
    3e.) Packages listed in "packages" are installed in @world.
    3f.) All files in scripts.postinst/ ending in .sh are executed.

Steps 2, 3a, 3c, 3e, and 3f are all optional.  Also note that the
container is run with user namespace support active (i.e., it uses the
"-U" option to systemd-nspawn).  Finally, the *.sh files must be shell
scripts.