Add a script that creates a GRUB image for EFI32
Add a script that creates a GRUB image for EFI32, which was useful on
diefledermaus (my old Mac Mini) for initially booting the system in EFI
mode using a pre-existing boot loader (e.g., rEFInd).  Basically, it is
used when grub-install cannot be, which is what usually generates the
EFI GRUB image automatically, but which cannot do so when the system was
booted in BIOS compatibility mode.

I'm mainly committing it because I don't want to forget how to do that.