rescan_block_dev: new script for resetting block devices
Add a new shell script rescan_block_dev for (soft-)resetting block
devices.  It is used in situations where a drive is not responding.  The
alternative would be to unplug it and plug it back in, but that loses
unwritten data in the SCSI cache (of the Linux kernel, that is).